split graphic DrawTarget implementations based on batch feature
